Fix getPartialResults(ISSUE).map()
Fix context.getPartialResults(ISSUE).map() so that it returns the
LintMap for the current project. Also fix writePartialResults so that it
writes the LintMap for the specified project. Add a test case that
previously failed.

Also changed LintCliClient's partial result HashMaps to LinkedHashMaps
so that iteration order of PartialResult is deterministic (insertion
order). This is not necessary for the fix, but can be helpful for users
when debugging issues with partial results.

Also fix VectorDrawableCompatDetector to search partial results of
dependent projects.

Bug: b/239337003
Test: Added test
Change-Id: I11918bd314e5f927830deb5e4cc450fadd01dab8