adblib: Make packet ID optional in JDWP session
There are cases where a consumer of a JDWPSession
does not need to send custom JDWP packets, hence
does not need to use the `JdwpSession.nextPacketId`
methods.

Before this change, such consumers were asked to
use a "random" value for `nextPacketIdBase`, and
`nextPacketId` would generate values based on that
"random" value .

With this change, consumers send a `null` value
for `nextPacketIdBase` and `nextPacketId` throws an
`UnsupportedOperationException` exception when
called. This makes for a more deterministic and
explicit behavior.

Bug: 188067646
Test: Included
Change-Id: Ia4e5c61270b0b0fef7165237161f83e69ad47535