Separate GradleVersion and AgpVersion
Implement (somewhat stricter) parsing of AGP versions: we now enforce
the two-digit format for alpha/beta/rc sub-versions for versions in
the range where that applies (essentially everything after 3.0.0, with
one surprising exception).  Also implement the various forms of
comparisons that currently exist (compareTo,
compareIgnoringQualifiers, isAtLeast, isAtLeastIncludingPreviews).

Keep AgpVersion in the GradleVersion namespace for now to avoid noise
in this and the accompanying change.

Bug: 242691473
Test: added tests
Change-Id: Ifa3c5c4f57870a08e6bec6ac8de6a43e0d273572