adblib: move "Shell" classes around
This change is a refactoring, there is no code behavior change.

* Move the base "shell" collector interfaces to the "AdbDeviceServices"
source file, since they are initially referenced there.

* Move all "shell collector" implementation classes to
the "ShellCommand" file, since they are referenced there
in the "withXxxCollector" extension functions. This makes it
more consistent, as before some implementation classes where
in the `adblib` package and some in the `adblib.utils` package.
They are now all part of the `adblib` package.

* Also move a couple of implementation classes to the `adblib.impl`
package, also for consistency.

Bug: 188067646
Test: Existing
Change-Id: I767531cd965805806d1acbe1b91cb3185992d72c