Add test for adding task output directory as a source set
This change was originally meant to fix a bug being hit in the AndroidX
build, but the new test passes with or without the changes to
AndroidLintInputs, so the fix for AndroidX is to add source sets there
as they're added in the new test.

Note: originally, I wrote the test more in line with how AndroidX
currently was adding their source set in
:room:room-compiler-processing-testing, but the change to
AndroidLintInputs didn't cause the test to pass, which makes sense
because I wouldn't expect testPropsOutDir to carry the desired
dependency information.

Bug: 253219347
Test: LintStandaloneTest
Change-Id: I549550cd134c180a9e18754f6b84d3184b292b43