Create drop-in replacement for unstable Kotlin Time API.
This API is experimental meaning the underlying implementation may
change, breaking existing code. This happens to us because our
runtime Kotlin may be different than our compile-time Kotlin.

This drop-in library attempts to provide basic compatible
functionality so that at a later date we can drop in the replacement
when it no longer requires the risk of an experimental API.

Bug: 254530383
Test: Included
Change-Id: Ib7c1bc9ca79b67370801c334975a1db8e7dbaf9a