Clean up: Make updating AndroidAttributes safer
Previously, we were making a copy of `AndroidAttributes` at the call
site. That means when the `AndroidAttributes` class evolves (e.g.,
having new properties), that change may not be reflected at the call
site.

In this commit, we will provide the `copy` method within the
`AndroidAttributes` class so it can evolve together with that class.

Bug: N/A (Clean up)
Test: Existing tests
Change-Id: If43276570f033722f3f31e2558c6b423e3733bf3