New Version class
Represent specific versions as interpreted by Gradle, and
(in future) as a building block for version specifiers, including rich
versions.  This will (again in future) replace all uses of
GradleVersion, GradleVersionRange and the version-related parts of
GradleCoordinate.

In the spirit of the classic idea that every service is either
deprecated or not ready yet, mark the new class as Beta.

Bug: 242691473
Test: added
Change-Id: Id6e47e215bc522c14ffe2c73deba46b31ba47a02