adblib: Add JdwpProcessTracker to device cache
Currently, only the ddmlib compatibility layer is instantiating
the `JdwpProcessTracker`. In the future, there will be additional
clients, and it would be more more efficient to have a single
`JdwpProcessTracker` instance per device, so we expose
a new per-device cache entry to store a re-usable
`JdwpProcessTracker`.

The ddmlib compatibility layer is updated to use this new
cache entry instead of creating a custom instance.

Bug: 188067646
Test: Existing
Change-Id: I68e49c57c720379041d0a5a12ebd5a1507d6206b